Move RTK offline modules into subfolder
