+++ rtk/src/index.ts
@@ -5,20 +5,20 @@ export * from "./betterAuthTypes";
 export * from "./constants";
 export * from "./emptyApi";
 export * from "./mongooseSlice";
-export * from "./offlineConfig";
-export * from "./offlineGate";
-export * from "./offlineIds";
-export * from "./offlineMiddleware";
-export * from "./offlineOptimistic";
-export * from "./offlineSlice";
-export * from "./offlineTypes";
+export * from "./offline/offlineConfig";
+export * from "./offline/offlineGate";
+export * from "./offline/offlineIds";
+export * from "./offline/offlineMiddleware";
+export * from "./offline/offlineOptimistic";
+export * from "./offline/offlineSlice";
+export * from "./offline/offlineTypes";
 export * from "./platform";
 export * from "./realtime";
 export * from "./socket";
 export * from "./sync";
 export * from "./tagGenerator";
 export * from "./useFeatureFlags";
-export * from "./useOfflineStatus";
+export * from "./offline/useOfflineStatus";
 export * from "./useRealtimeDebug";
-export * from "./useServerStatus";
+export * from "./offline/useServerStatus";
 export * from "./useUpgradeCheck";
